+++ src/main/java/com/arcbees/bourseje/client/application/voteinactive/VoteInactiveView.ui.xml
